add reload test for behat
- Modify FeatureContext in order to test locally without dump.
